+++ packages/preview-middleware-client/src/adp/controllers/AddFragment.controller.ts
@@ -292,6 +292,8 @@ export default class AddFragment extends BaseDialog<AddFragmentModel> {
             targetAggregation === 'columns'
         ) {
             return 'ANALYTICAL_TABLE_COLUMN';
+        } else if (currentControlName === 'sap.ui.mdc.ActionToolbar' && targetAggregation === 'actions') {
+            return 'TABLE_ACTION';
         }
         return '';
     }