+++ packages/preview-middleware-client/src/adp/quick-actions/fe-v4/registry.ts
@@ -10,6 +10,8 @@ import { ChangeTableColumnsQuickAction } from './change-table-columns';
 import { AddHeaderFieldQuickAction } from '../common/op-add-header-field';
 import { AddCustomSectionQuickAction } from '../common/op-add-custom-section';
 import { AddTableCustomColumnQuickAction } from './create-table-custom-column';
+import { AddPageActionQuickAction } from '../common/create-page-action';
+import { AddTableActionQuickAction } from './create-table-action';
 
 type PageName = 'listReport' | 'objectPage';
 
@@ -24,13 +26,12 @@ export default class FEV4QuickActionRegistry extends QuickActionDefinitionRegist
         [LIST_REPORT_TYPE]: 'listReport',
         [OBJECT_PAGE_TYPE]: 'objectPage'
     };
+
     getDefinitions(context: QuickActionActivationContext): QuickActionDefinitionGroup[] {
         const activePages = this.getActivePageContent(context.controlIndex);
-
         const definitionGroups: QuickActionDefinitionGroup[] = [];
         for (let index = 0; index < activePages.length; index++) {
             const { name, view } = activePages[index];
-
             if (name === 'listReport') {
                 definitionGroups.push({
                     title: 'LIST REPORT',
@@ -38,6 +39,8 @@ export default class FEV4QuickActionRegistry extends QuickActionDefinitionRegist
                         ToggleClearFilterBarQuickAction,
                         AddControllerToPageQuickAction,
                         ChangeTableColumnsQuickAction,
+                        AddPageActionQuickAction,
+                        AddTableActionQuickAction,
                         AddTableCustomColumnQuickAction
                     ],
                     view,
@@ -51,6 +54,8 @@ export default class FEV4QuickActionRegistry extends QuickActionDefinitionRegist
                         ChangeTableColumnsQuickAction,
                         AddHeaderFieldQuickAction,
                         AddCustomSectionQuickAction,
+                        AddPageActionQuickAction,
+                        AddTableActionQuickAction,
                         AddTableCustomColumnQuickAction
                     ],
                     view,